requirements.txt
motor~=3.4.0
beanie~=1.26.0
pydantic~=2.7.3
datasketch~=1.6.5
pandas~=2.2.2
requests~=2.32.3
PyYAML~=6.0.1
beautifulsoup4~=4.12.3